sprint 3/20/18
i am going to focus of project euler for a little while.
the interface is working as planed in project_euler.py, but the format i wrote problem_1() as is not to my liking.

eulerProblemTest is the document im going to be using to test each problem individualy and isolated from the rest of the code.